construct authenticateJWT function as middleware
