+++ app.js
@@ -15,6 +15,8 @@ const userRoutes = require("./routes/userRoutes");
 
 const port = process.env.PORT
 
+const authenticateJWT = require('./middleware/authMiddleware')
+
 app.use(express.json());
 app.use(bodyParser.json());
 
@@ -53,31 +55,6 @@ app.post('/login', async (req, res) => {
 
     res.json({ token });
 });
-
-const authenticateJWT = (req, res, next) => {
-    const authHeader = req.header('Authorization');
-    
-    if (!authHeader) {
-        console.log('No Authorization header');
-        return res.status(401).send('Access denied');
-    }
-
-    const token = authHeader.split(' ')[1];
-
-    if (!token) {
-        console.log('No token in header');
-        return res.status(401).send('Access denied');
-    }
-
-    try {
-        const decoded = jwt.verify(token, secretKey);
-        req.user = decoded;
-        next();
-    } catch (err) {
-        console.log('Token verification failed', err.message);
-        res.status(400).send('Invalid token');
-    }
-};
 
 app.get('/protected', authenticateJWT, (req, res) => {
     res.send('This is a protected route');
